+++ nrfutil/device/device.ts
@@ -6,10 +6,12 @@
 
 import { LogLevel, LogMessage } from '../sandboxTypes';
 import { Batch } from './batch';
+import boardController from './boardController';
 import { getDeviceSandbox } from './common';
 import deviceInfo from './deviceInfo';
 import erase from './erase';
 import firmwareRead from './firmwareRead';
+import getBoardControllerVersion from './getBoardControllerVersion';
 import getCoreInfo from './getCoreInfo';
 import getFwInfo from './getFwInfo';
 import getProtectionStatus from './getProtectionStatus';
@@ -59,5 +61,7 @@ export default {
     setLogLevel,
     setVerboseLogging,
     getModuleVersion,
+    boardController,
+    getBoardControllerVersion,
     batch: () => new Batch(),
 };